src/main/java/com/springboot/aws/project/web/HelloController.java
package com.springboot.aws.project.web;

import com.springboot.aws.project.web.dto.HelloResponseDto;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController //JSON반환
public class HelloController {

    @GetMapping("/hello")
    public String hello(){
        return "hello"; //@RequestMapping
    }

    @GetMapping("/hello/dto")
    public HelloResponseDto helloDto(@RequestParam("name") String name, @RequestParam("amount") int amount){
        return new HelloResponseDto(name, amount);
    }
}
